WordCamp: Try casting `error_reporting` result as an integer
This should already be an integer, but it's not getting caught correctly
in our custom error handler.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8261 74240141-8908-4e6f-9713-ba540dce6ec7

+++ public_html/wp-content/mu-plugins/wcorg-misc.php
@@ -467,7 +467,7 @@ function send_error_to_slack( $err_no, $err_msg, $file, $line ) {
 
 	// Checks to see if the error-throwing expression is prepended with the @ control operator.
 	// See https://secure.php.net/manual/en/function.set-error-handler.php
-	if ( 0 === error_reporting() ) {
+	if ( 0 === (int) error_reporting() ) {
 		return false;
 	}
